Split background script in smaller parts
