+++ webpack.config.js
@@ -3,7 +3,7 @@ const CopyWebpackPlugin = require('copy-webpack-plugin'); // eslint-disable-line
 module.exports = {
   entry: {
     app: './lib/app',
-    background: './lib/background',
+    background: './lib/background/index',
     options: './lib/options/page',
   },
   devtool: 'source-map',
